Refactored cmd logic. Added command to list test definitions
