+++ pkg/kyma/cmd/test/common_test.go
@@ -10,66 +10,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func Test_ListTestDefinitionNames(t *testing.T) {
-	testData := []struct {
-		testName         string
-		shouldFail       bool
-		inputDefinitions oct.TestDefinitionList
-		expectedResult   []string
-	}{
-		{
-			testName:   "correct list",
-			shouldFail: false,
-			inputDefinitions: oct.TestDefinitionList{
-				Items: []oct.TestDefinition{
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "test1",
-						},
-					},
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "test2",
-						},
-					},
-				},
-			},
-			expectedResult: []string{"test1", "test2"},
-		},
-		{
-			testName:   "incorrect list",
-			shouldFail: true,
-			inputDefinitions: oct.TestDefinitionList{
-				Items: []oct.TestDefinition{
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "test1",
-						},
-					},
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "test2",
-						},
-					},
-				},
-			},
-			expectedResult: []string{"test1", "test3"},
-		},
-	}
-
-	for _, tt := range testData {
-		mCli := octopus.NewMockedOctopusRestClient(&tt.inputDefinitions, nil)
-		dNames, err := ListTestDefinitionNames(mCli)
-		if !tt.shouldFail {
-			require.Nil(t, err, tt.testName)
-			require.True(t, reflect.DeepEqual(dNames, tt.expectedResult))
-		} else {
-			require.False(t, reflect.DeepEqual(dNames, tt.expectedResult))
-		}
-
-	}
-}
-
 func Test_ListTestSuiteNames(t *testing.T) {
 	testData := []struct {
 		testName        string
