+++ pkg/kyma/cmd/test/definitions/opts.go
@@ -1,4 +1,4 @@
-package list
+package definitions
 
 import "github.com/kyma-project/cli/pkg/kyma/core"
