+++ pkg/kyma/cmd/test/run/opts.go
@@ -1,6 +1,8 @@
 package run
 
 import (
+	"time"
+
 	"github.com/kyma-project/cli/pkg/kyma/core"
 )
 
@@ -8,7 +10,7 @@ type options struct {
 	*core.Options
 	Name    string
 	Wait    bool
-	Timeout int
+	Timeout time.Duration
 }
 
 func NewOptions(o *core.Options) *options {
